requirements.txt
Adafruit-Blinka==8.66.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-motor==3.4.17
adafruit-circuitpython-pca9685==3.4.19
adafruit-circuitpython-register==1.10.4
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-servokit==1.3.21
adafruit-circuitpython-typing==1.12.2
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
Adafruit-SSD1306==1.6.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
babel==2.17.0
bcrypt==3.2.0
beautifulsoup4==4.13.5
beniget==0.4.1
bidict==0.23.1
binho-host-adapter==0.1.6
bleach==6.2.0
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
certifi==2025.11.12
cffi==2.0.0
chardet==4.0.0
charset-normalizer==3.4.3
click==8.0.3
colorama==0.4.4
comm==0.2.3
cpuset==1.6
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.17
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.2
docutils==0.22.3
duplicity==0.8.21
exceptiongroup==1.3.0
executing==2.2.1
-e git+https://github.com/FaBoPlatform/FaBoPWM-PCA9685-Python@5a89518bba5721a30b43bd4bd2d64a39286d169c#egg=Fabo_PCA9685
fastapi==0.121.3
fasteners==0.14.1
fastjsonschema==2.21.2
filelock==3.19.1
fonttools==4.29.1
frozenlist==1.8.0
fs==2.4.12
fsspec==2025.9.0
future==0.18.2
gast==0.5.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.3
ifaddr==0.2.0
ImageIO==2.37.2
importlib-metadata==4.6.4
ipykernel==6.30.1
ipython==8.37.0
ipython-genutils==0.2.0
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
jetcam @ git+https://github.com/NVIDIA-AI-IOT/jetcam@508ff3a402c5a0449a8d5776c05d2d17d6a87f46
-e git+https://github.com/FaBoPlatform/jetracer@069e5f9870e4a33745084f516c2c9fe45a6c4476#egg=jetracer
Jetson.GPIO==2.1.7
Jinja2==3.1.6
json5==0.12.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
-e git+https://github.com/tokk-nv/jupyter_clickable_image_widget@b3b38ebe1dbeb5bbf0df52034ea4e58ff603589e#egg=jupyter_clickable_image_widget
jupyter-console==6.6.3
jupyter-server==1.24.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab==3.2.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
llama_cpp_python==0.3.16
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
markdown2==2.5.4
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.7
mcp==1.23.1
mistune==3.1.4
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
multidict==6.7.0
nbclassic==0.5.6
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nicegui==3.3.1
notebook==6.5.4
notebook_shim==0.2.4
numpy==2.2.6
oauthlib==3.2.0
olefile==0.46
onboard==1.4.1
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==1.3.5
pandocfilters==1.5.1
paramiko==2.9.3
parso==0.8.5
pbr==7.0.1
pexpect==4.8.0
pillow==12.0.0
platformdirs==4.4.0
ply==3.11
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.12.4
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycparser==2.23
pycups==2.0.1
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyftdi==0.57.1
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.10.1
pymacaroons==0.13.0
PyNaCl==1.5.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyRFC3339==1.1
pyserial==3.5
pyservicemaker @ file:///opt/nvidia/deepstream/deepstream-7.1/service-maker/python/pyservicemaker-0.0.1-py3-none-linux_aarch64.whl
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-dbusmock==0.27.5
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.2.1
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.15.0
pythran==0.10.0
pytz==2022.7.1
pyusb==1.3.1
pyxdg==0.27
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
scikit-image==0.25.2
scipy==1.15.3
SecretStorage==3.3.1
Send2Trash==1.8.3
simple-websocket==1.1.0
six==1.16.0
smbus==1.1.post2
sniffio==1.3.1
soupsieve==2.8
spidev==3.7
sse-starlette==3.0.3
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
systemd-python==234
sysv-ipc==1.1.0
tensorrt==10.7.0
tensorrt_dispatch==10.7.0
tensorrt_lean==10.7.0
terminado==0.18.1
testresources==2.0.1
tifffile==2025.5.10
tinycss2==1.4.0
toml==0.10.2
torch @ file:///home/jetson/Jetson_script/jetracer/orin_nano/6.2.1/jetracer/torch-2.8.0-cp310-cp310-linux_aarch64.whl
torch2trt==0.5.0
torchvision @ file:///home/jetson/Jetson_script/jetracer/orin_nano/6.2.1/jetracer/torchvision-0.23.0-cp310-cp310-linux_aarch64.whl
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
unicodedata2==14.0.0
urllib3==1.26.5
urwid==2.1.2
uvicorn==0.38.0
uvloop==0.22.1
wadllib==1.3.6
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
widgetsnbextension==4.0.14
wsproto==1.3.2
xdg==5
xkit==0.0.0
yarl==1.22.0
zipp==1.0.0
